# spec/openapi/drone-backend.openapi.json
{
  "components" : {
    "schemas" : {
...
      "GpsPointVO" : {
        "properties" : {
          "altitude" : {
            "format" : "double",
            "type" : "number"
          },
          "battery" : {
            "format" : "int32",
            "type" : "integer"
          },
          "flightStatus" : {
            "format" : "int32",
            "type" : "integer"
          },
          "latitude" : {
            "format" : "double",
            "type" : "number"
          },
          "longitude" : {
            "format" : "double",
            "type" : "number"
          },
          "operation" : {
            "type" : "string"
          },
          "speed" : {
            "format" : "double",
            "type" : "number"
          },
          "timestamp" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RecordItem" : {
        "description" : "单条观看记录",
        "properties" : {
          "djiId" : {
            "description" : "设备ID(DJI ID)",
            "type" : "string"
          },
          "endTime" : {
            "description" : "观看结束时间（ISO-8601；未结束时为 null）",
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "description" : "记录ID",
            "format" : "int64",
            "type" : "integer"
          },
          "startTime" : {
            "description" : "观看开始时间（ISO-8601，如 2026-05-31T10:00:00）",
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "RegisterVo" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "userName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ResultGpsPointVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/GpsPointVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultListGpsPointVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）",
            "items" : {
              "$ref" : "#/components/schemas/GpsPointVO"
            },
            "type" : "array"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultListUavVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）",
            "items" : {
              "$ref" : "#/components/schemas/UavVo"
            },
            "type" : "array"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultMapStringObject" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "additionalProperties" : { },
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）",
            "type" : "object"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultRegisterVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/RegisterVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultTokenRefreshVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/TokenRefreshVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultUserLoginVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/UserLoginVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ResultUserRecordsVO" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/UserRecordsVO",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ResultWebUavStatusVo" : {
        "description" : "统一响应结果封装",
        "properties" : {
          "code" : {
            "description" : "响应状态码（200 成功；400 参数或业务错误；401 未登录；404 资源不存在；409 状态冲突；500 服务端错误）",
            "format" : "int32",
            "type" : "integer"
          },
          "data" : {
            "$ref" : "#/components/schemas/WebUavStatusVo",
            "description" : "业务数据（泛型；成功且有返回值时非空，无数据时为 null）"
          },
          "errorCode" : {
            "description" : "业务错误码（失败时非空，取值见 ApiErrorCode 枚举，如 UAV_NOT_CONNECTED；成功时为 null）",
            "type" : "string"
          },
          "message" : {
            "description" : "提示信息（成功为“操作成功”，失败为具体失败原因）",
            "type" : "string"
          },
          "success" : {
            "description" : "是否成功：true 成功 / false 失败",
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "TokenRefreshVO" : {
        "properties" : {
          "token" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UavVo" : {
        "properties" : {
          "controllerModel" : {
            "description" : "控制器型号",
            "type" : "string"
          },
          "djiId" : {
            "description" : "设备ID(DJI ID)",
            "type" : "string"
          },
          "id" : {
            "description" : "id",
            "format" : "int64",
            "type" : "integer"
          },
          "isAvailable" : {
            "description" : "是否可用(1可用,0禁用)",
            "type" : "string"
          },
          "uavName" : {
            "description" : "uavName",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UserLoginDto" : {
        "description" : "Parameters for Login Request(登录参数)",
        "properties" : {
          "password" : {
            "description" : "password",
            "type" : "string"
          },
          "userName" : {
            "description" : "userName",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UserLoginVO" : {
        "properties" : {
          "refreshToken" : {
            "type" : "string"
          },
          "token" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UserRecordsVO" : {
        "description" : "用户观看记录分页结果",
        "properties" : {
          "records" : {
            "description" : "当前页观看记录列表",
            "items" : {
              "$ref" : "#/components/schemas/RecordItem"
            },
            "type" : "array"
          },
          "total" : {
            "description" : "观看记录总条数",
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "description" : "总页数",
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "UserRegisterDto" : {
        "description" : "Parameters for Register Request(注册参数)",
        "properties" : {
          "password" : {
            "description" : "passWord",
            "type" : "string"
          },
          "userName" : {
            "description" : "userName",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "WebUavStatusVo" : {
        "description" : "Web 端无人机状态（含连接状态与直播状态）",
        "properties" : {
          "djiId" : {
            "description" : "设备ID(DJI ID)",
            "type" : "string"
          },
          "id" : {
            "description" : "无人机ID",
            "format" : "int64",
            "type" : "integer"
          },
          "latestStatus" : {
            "$ref" : "#/components/schemas/UavRuntimeStatusVo",
            "description" : "最近一次上报的运行状态（设备暂未上报时为 null）"
          },
          "liveState" : {
            "description" : "直播会话状态（IDLE 空闲 / STARTING 启动中 / RUNNING 运行中）",
            "type" : "string"
          },
          "uavName" : {
            "description" : "无人机名称",
            "type" : "string"
          },
          "wsConnected" : {
            "description" : "设备 WebSocket 是否已连接",
            "type" : "boolean"
          }
        },
        "type" : "object"
      }
    },
    "securitySchemes" : {
      "Bearer Authentication" : {
        "bearerFormat" : "JWT",
        "description" : "请输入JWT Token，格式：直接输入token字符串即可",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "无人机项目web后端接口文档",
    "title" : "无人机项目接口文档",
    "version" : "1.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/user/login" : {
      "post" : {
        "description" : "验证用户名和密码，成功返回 JWT 令牌",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserLoginDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultUserLoginVO"
                }
              }
            },
            "description" : "登录成功"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultUserLoginVO"
                }
              }
            },
            "description" : "用户名或密码错误"
          }
        },
        "summary" : "用户登录",
        "tags" : [ "Authentication and Authorization API" ]
      }
    },
    "/user/records" : {
      "get" : {
        "description" : "获取当前登录用户的直播观看记录，支持分页",
        "operationId" : "getLiveRecords",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 10,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultUserRecordsVO"
                }
              }
            },
            "description" : "查询成功"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultUserRecordsVO"
                }
              }
            },
            "description" : "未登录"
          }
        },
        "summary" : "查询用户直播记录",
        "tags" : [ "Authentication and Authorization API" ]
      }
    },
    "/user/refresh" : {
      "post" : {
        "description" : "使用 Refresh-Token 获取新的 Access-Token",
        "operationId" : "refreshToken",
        "parameters" : [ {
          "in" : "header",
          "name" : "Refresh-Token",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTokenRefreshVO"
                }
              }
            },
            "description" : "刷新成功"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultTokenRefreshVO"
                }
              }
            },
            "description" : "刷新失败"
          }
        },
        "summary" : "刷新令牌",
        "tags" : [ "Authentication and Authorization API" ]
      }
    },
    "/user/register" : {
      "post" : {
        "description" : "用户注册，返回用户信息",
        "operationId" : "registerUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserRegisterDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultRegisterVo"
                }
              }
            },
            "description" : "注册成功"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultRegisterVo"
                }
              }
            },
            "description" : "注册失败"
          }
        },
        "summary" : "用户注册",
        "tags" : [ "Authentication and Authorization API" ]
      }
    },
    "/webUav/getRecord" : {
      "get" : {
        "description" : "普通用户固定查询自己的直播观看记录；管理员可传 userName 代查任意用户",
        "operationId" : "getUserRecord",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 10,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "userName",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultUserRecordsVO"
                }
              }
            },
            "description" : "查询成功"
          }
        },
        "summary" : "查询观看记录（本人）",
        "tags" : [ "WebUav API" ]
      }
    },
    "/webUav/getUav" : {
      "get" : {
        "description" : "获取所有的无人机列表",
        "operationId" : "getUav",
        "parameters" : [ {
          "in" : "query",
          "name" : "onlineOnly",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListUavVo"
                }
              }
            },
            "description" : "查询成功"
          }
        },
        "summary" : "查询无人机",
        "tags" : [ "WebUav API" ]
      }
    },
    "/webUav/position" : {
      "get" : {
        "description" : "优先内存实时数据，内存无数据时降级为 DB 最近 2 分钟内的记录",
        "operationId" : "getPosition",
        "parameters" : [ {
          "description" : "无人机 ID（与 deviceId 二选一）",
          "in" : "query",
          "name" : "uavId",
          "required" : false,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "description" : "设备 ID（与 uavId 二选一）",
          "in" : "query",
          "name" : "deviceId",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultGpsPointVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "查询无人机实时位置",
        "tags" : [ "WebUav API" ]
      }
    },
    "/webUav/reverseGeocode" : {
      "get" : {
        "description" : "调用高德逆地理编码 API，将经纬度转为格式化地址（结果缓存 4 位小数精度）",
        "operationId" : "reverseGeocode",
        "parameters" : [ {
          "description" : "纬度",
          "in" : "query",
          "name" : "lat",
          "required" : true,
          "schema" : {
            "format" : "double",
            "type" : "number"
          }
        }, {
          "description" : "经度",
          "in" : "query",
          "name" : "lng",
          "required" : true,
          "schema" : {
            "format" : "double",
            "type" : "number"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultMapStringObject"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "坐标转地址",
        "tags" : [ "WebUav API" ]
      }
    },
    "/webUav/status" : {
      "get" : {
        "description" : "根据设备ID查询平台侧最近一次收到的无人机状态",
        "operationId" : "getUavStatus",
        "parameters" : [ {
          "in" : "query",
          "name" : "deviceId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultWebUavStatusVo"
                }
              }
            },
            "description" : "查询成功"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultWebUavStatusVo"
                }
              }
            },
            "description" : "无人机或状态不存在"
          }
        },
        "summary" : "查询单台无人机实时状态",
        "tags" : [ "WebUav API" ]
      }
    },
    "/webUav/trajectory" : {
      "get" : {
        "description" : "根据订单号查询无人机执行该订单时的 GPS 轨迹，按时间升序排列",
        "operationId" : "getTrajectory",
        "parameters" : [ {
          "description" : "订单号",
          "in" : "query",
          "name" : "orderNum",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ResultListGpsPointVO"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "查询订单飞行轨迹",
        "tags" : [ "WebUav API" ]
      }
    }
  },
  "security" : [ {
    "Bearer Authentication" : [ ]
  } ],
  "servers" : [ {
    "description" : "相对基址：由部署环境（网关/反向代理）决定",
    "url" : "/"
  } ],
...
}